create model_utils.py
