--- test/README.md
@@ -1,13 +1,0 @@
-# Tests
-
-The current code is just a placeholder, need to determine how to:
-
-- simplify model setup
-- appropriate smaller-scale computations
-- consistent random seeding to obtain deterministic checkable results.
-
-Example:
-
-```
-pytest test_rbm.py
-````
